feat: integrate memory page with read-only api
Mugiwara-Agent: zoro
